refac: 개인정보처리방침과 서비스 이용약관 모달
