+++ src/assets/terms-and-policy/policyOfHandlingPersonalInformation.html
@@ -383,7 +383,6 @@
         ><span>&#xa0;</span
         ><a
           href="https://developers.google.com/terms/api-services-user-data-policy"
-          style="text-decoration: none"
           target="_blank"
           ><span class="Hyperlink"
             >Google API 서비스 사용자 데이터 정책</span
@@ -1476,46 +1475,31 @@
       <ol type="1" style="margin: 0pt; padding-left: 0pt">
         <li style="margin-left: 31.35pt; padding-left: 4.65pt">
           <span>개인정보침해신고센터 (</span
-          ><a
-            href="http://www.1336.or.kr/"
-            style="text-decoration: none"
-            target="_blank"
+          ><a href="http://www.1336.or.kr/" target="_blank"
             ><span class="Hyperlink">www.1336.or.kr</span></a
           ><span>&#xa0;</span><span>/ 국번없이 118)</span>
         </li>
         <li style="margin-left: 31.35pt; padding-left: 4.65pt">
           <span>정보보호마크인증위원회 (</span
-          ><a
-            href="http://www.eprivacy.or.kr/"
-            style="text-decoration: none"
-            target="_blank"
+          ><a href="http://www.eprivacy.or.kr/" target="_blank"
             ><span class="Hyperlink">www.eprivacy.or.kr</span></a
           ><span>&#xa0;</span><span>/ 02-580-0533~4)</span>
         </li>
         <li style="margin-left: 31.35pt; padding-left: 4.65pt">
           <span>대검찰청 사이버범죄수사센터 (</span
-          ><a
-            href="http://www.spo.go.kr/minwon"
-            style="text-decoration: none"
-            target="_blank"
+          ><a href="http://www.spo.go.kr/minwon" target="_blank"
             ><span class="Hyperlink">www.spo.go.kr/minwon</span></a
           ><span>&#xa0;</span><span>/ 국번 없이 1301)</span>
         </li>
         <li style="margin-left: 31.35pt; padding-left: 4.65pt">
           <span>경찰청 사이버테러대응센터 (</span
-          ><a
-            href="http://www.ctrc.go.kr/"
-            style="text-decoration: none"
-            target="_blank"
+          ><a href="http://www.ctrc.go.kr/" target="_blank"
             ><span class="Hyperlink">www.ctrc.go.kr</span></a
           ><span>&#xa0;</span><span>/ 02-392-0330)</span>
         </li>
         <li style="margin-left: 31.35pt; padding-left: 4.65pt">
           <span>개인정보분쟁조정위원회 (</span
-          ><a
-            href="http://cyberbureau.police.go.kr/"
-            style="text-decoration: none"
-            target="_blank"
+          ><a href="http://cyberbureau.police.go.kr/" target="_blank"
             ><span class="Hyperlink">cyberbureau.police.go.kr</span></a
           ><span>&#xa0;</span><span>/ 1833-6972)</span>
         </li>
