+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useStoreMap } from '@/hooks/useStoreMap';
 export { useCollapsibleSideView } from '@/hooks/useCollapsibleSideView';
 export { useToggleModal } from '@/hooks/useToggleModal';
+export { useTermsAndPolicy } from '@/hooks/useTermsAndPolicy';
